pyproject.toml
[project]
name = "riddle-llm-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "cfiken", email = "kentaro.nakanishi@abejainc.com" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.5",
    "google-generativeai>=0.8.5",
    "litellm>=1.80.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]

[project.scripts]
riddle-benchmark = "riddle_benchmark.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/riddle_benchmark"]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]

# --- Tool Configurations ---

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
